fix: return type for client upload
