+++ packages/storage/src/server.ts
@@ -1,6 +1,10 @@
-export * from './lib/get';
-export * from './lib/list';
-export * from './lib/head';
-export * from './lib/remove';
-export * from './lib/put';
-export * from './lib/presigned-url';
+export { type GetOptions, type GetResponse, get } from './lib/get';
+export { type ListOptions, type ListResponse, list } from './lib/list';
+export { type HeadOptions, type HeadResponse, head } from './lib/head';
+export { type RemoveOptions, remove } from './lib/remove';
+export { type PutOptions, type PutResponse, put } from './lib/put';
+export {
+  type GetPresignedUrlOptions,
+  type GetPresignedUrlResponse,
+  getPresignedUrl,
+} from './lib/presigned-url';
